keras/detection/new_0826/SSD_ipynb_transfer_callback_1227hand/requirements.txt
beautifulsoup4 
Cython
dm-tree==0.1.8
flatbuffers==23.1.21
google-auth==2.16.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
h5py==3.8.0
ipykernel
jupyter
keras==2.11.0
Markdown==3.4.1
markdown-it-py==3.0.0
matplotlib==3.7.2
mkl-fft==1.3.1
mkl-service==2.4.0
model-compression-toolkit==1.8.0
notebook
numpy
onnx==1.13.0
opencv-python==4.7.0.72
pandas==1.5.3
Pillow==9.2.0
protobuf==3.20.3
pycocotools==2.0
PyQt5==5.15.9
PyYAML==6.0
requests 
rich==13.5.2
# scikit-image==0.20.0
# scikit-learn==1.2.2
scipy==1.10.1
# tensorboard==2.11.2
# tensorflow==2.11.0
# torch
# torchvision 
tornado
tqdm==4.65.0